requirements.txt
fastapi==0.95.1
uvicorn==0.22.0
websockets==11.0.3
python-multipart==0.0.6
jinja2==3.1.2
aiohttp==3.8.4
PyPDF2==3.0.1
faiss-cpu==1.7.4
sentence-transformers==2.2.2
scikit-learn==1.2.2
numpy==1.24.3
python-dotenv==1.0.0
asyncio==3.4.3
typing-extensions==4.5.0
pydantic==1.10.7